responsive home page
